Add new configuration interface hierarchy to model both "portable" and "rich" configurations.
A "portable" configuration is a representation of a Dogma application's
configuration that does not depend on the underlying Go types it uses. Hence,
it can be serialized and transmitted across the network, even to Go processes
that do not have access to the application's Go types.

A "rich" configuration is one that *does* have access to the Go types. It is
equivalent to the implementation from `github.com/dogmatiq/enginekit/config`
that serves as the basis for this module.